Separate code from ChatMessageV2
As long as client-side code is only importing types from schema files,
we can potentially try and strip out the io-ts code from making it to
the client. To do that, we need to remove code from schemas that's
imported elsewhere.
