+++ imports/client/components/FirehosePage.tsx
@@ -15,7 +15,8 @@ import { indexedById } from '../../lib/listUtils';
 import ChatMessages from '../../lib/models/ChatMessages';
 import { indexedDisplayNames } from '../../lib/models/MeteorUsers';
 import Puzzles from '../../lib/models/Puzzles';
-import { ChatMessageType, nodeIsMention } from '../../lib/schemas/ChatMessage';
+import nodeIsMention from '../../lib/nodeIsMention';
+import { ChatMessageType } from '../../lib/schemas/ChatMessage';
 import { PuzzleType } from '../../lib/schemas/Puzzle';
 import { useBreadcrumb } from '../hooks/breadcrumb';
 import useSubscribeDisplayNames from '../hooks/useSubscribeDisplayNames';
