+++ imports/client/components/PuzzlePage.tsx
@@ -59,8 +59,10 @@ import Hunts from '../../lib/models/Hunts';
 import MeteorUsers, { indexedDisplayNames } from '../../lib/models/MeteorUsers';
 import Puzzles from '../../lib/models/Puzzles';
 import Tags from '../../lib/models/Tags';
+import nodeIsMention from '../../lib/nodeIsMention';
+import nodeIsText from '../../lib/nodeIsText';
 import { userMayWritePuzzlesForHunt } from '../../lib/permission_stubs';
-import { ChatMessageType, nodeIsMention, nodeIsText } from '../../lib/schemas/ChatMessage';
+import { ChatMessageType } from '../../lib/schemas/ChatMessage';
 import { DocumentType } from '../../lib/schemas/Document';
 import { GuessType } from '../../lib/schemas/Guess';
 import { PuzzleType } from '../../lib/schemas/Puzzle';
