+++ imports/server/hooks/DiscordHooks.ts
@@ -6,7 +6,8 @@ import MeteorUsers from '../../lib/models/MeteorUsers';
 import Puzzles from '../../lib/models/Puzzles';
 import Settings from '../../lib/models/Settings';
 import Tags from '../../lib/models/Tags';
-import { ChatMessageContentType, nodeIsText } from '../../lib/schemas/ChatMessage';
+import nodeIsText from '../../lib/nodeIsText';
+import { ChatMessageContentType } from '../../lib/schemas/ChatMessage';
 import { computeSolvedness } from '../../lib/solvedness';
 import { DiscordBot } from '../discord';
 import Hookset from './Hookset';
